+++ data/albums.json
@@ -1,1 +1,1 @@
-{"publish_date":"1234-11-11","id":"d666134c-2121-4a7f-8532-e8cd0f88d7d1","archived":false,"author":{"id":"ffca4b1e-c8f4-43ef-adc6-8b3f22199b0c","first_name":"Ubisoft","last_name":""},"genre":{"id":"6739afba-fe48-4c66-83fd-b8dbee114bf7","name":"Action"},"label":{"id":414,"title":null,"color":"Blue"},"on_spotify":false}
+{"publish_date":"1234-11-11","id":"d666134c-2121-4a7f-8532-e8cd0f88d7d1","archived":false,"author":{"id":"a68452ee-a07b-4967-9951-0d995c551e7e","first_name":"Ubisoft","last_name":""},"genre":{"id":"65c0318c-fb1c-4dd8-b885-06848e4c0c24","name":"Action"},"label":{"id":262,"title":null,"color":"Blue"},"on_spotify":false}
