+++ data/books.json
@@ -1,1 +1,1 @@
-{"publisher":"Me","cover_state":"good","publish_date":"1234-11-11","id":"a29e4867-8318-48c9-b9a7-5fa65125993f","archived":false,"author":{"id":"df362a90-0ea2-4432-a007-817c4ef80670","first_name":"Ubisoft","last_name":""},"genre":{"id":"5a600be1-d634-4060-b07e-b282a3ae1ae8","name":"Action"},"label":{"id":343,"title":null,"color":"Blue"}}
+{"publisher":"Me","cover_state":"good","publish_date":"1234-11-11","id":"a29e4867-8318-48c9-b9a7-5fa65125993f","archived":false,"author":{"id":"03dfd534-6419-4ae4-9894-83d52c3bc404","first_name":"Ubisoft","last_name":""},"genre":{"id":"666a7889-a305-445c-9129-256f2d48f9bc","name":"Action"},"label":{"id":48,"title":null,"color":"Blue"}}
