+++ data/games.json
@@ -1,1 +1,1 @@
-{"publish_date":"1234-11-11","archived":false,"author":{"id":"6abf5473-ef10-42a8-97d7-64d27a21eaf5","first_name":"Ubisoft","last_name":""},"genre":{"id":"085513c2-6cd5-4442-9ab0-bd31f2aaf5a5","name":"Action"},"label":{"id":281,"title":null,"color":"Blue"},"multiplayer":"n","last_played_at":"2023-02-02","id":"37bd485f-8991-4ce9-a911-23ed74edcbfc"}
+{"publish_date":"1234-11-11","archived":false,"author":{"id":"fa6511e0-1789-4d7f-ae81-86b8fd730e9f","first_name":"Ubisoft","last_name":""},"genre":{"id":"dbff3ab5-e146-4418-9ea9-a69b7c94d2a0","name":"Action"},"label":{"id":20,"title":null,"color":"Blue"},"multiplayer":"n","last_played_at":"2023-02-02","id":"37bd485f-8991-4ce9-a911-23ed74edcbfc"}
